refactor(step4): change package name - step3 -> racingcar
