+++ src/main/kotlin/racingcar/domain/Position.kt
@@ -1,4 +1,4 @@
-package step3.domain
+package racingcar.domain
 
 import kotlin.math.max
