+++ src/test/kotlin/racingcar/domain/PositionTest.kt
@@ -1,4 +1,4 @@
-package step3.domain
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
